requirements.txt
Flask==3.0.3
Jinja2==3.1.4
python-escpos==3.1
waitress==3.0.0
typer==0.12.3
python-dotenv==1.0.1

